add prompt functions for continue_conversation, build_context, recent_activity
