+++ src/basic_memory/mcp/prompts/__init__.py
@@ -7,4 +7,4 @@ and transforming their results into user-friendly text.
 
 # Import individual prompt modules to register them with the MCP server
 from basic_memory.mcp.prompts import guide
-from basic_memory.mcp.prompts import session
+from basic_memory.mcp.prompts import continue_conversation
